requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

langchain==0.3.3
langchain-core==0.3.10
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
langchain-google-genai==1.0.5

chromadb==0.5.5
sentence-transformers==3.0.1
torch==2.4.0
transformers==4.44.0
unstructured==0.15.0
pymupdf==1.24.8
pikepdf==9.2.0
